+++ monitoring/monitoring/monitoring_app.py
@@ -1,7 +1,7 @@
 import streamlit as st
 from database.connection import GreenDB
 
-from monitoring.render_helper import (
+from monitoring.utils import (
     render_basic_information,
     render_extended_information,
     render_sidebar,
